requirements.txt
aiogram
python-dotenv
gino
sqlalchemy